% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{checkForegrounds}
\alias{checkForegrounds}
\title{Subset foreground species based on existing species in alignment file}
\usage{
checkForegrounds(align, foregrounds)
}
\arguments{
\item{align}{alignment object read from read.msa}

\item{foregrounds}{character vector containing the foreground species}
}
\value{
foregrounds.out character vector containing the edited foregrounds
}
\description{
Subset foreground species based on existing species in alignment file
}
