% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{convertAlignmentToFeature}
\alias{convertAlignmentToFeature}
\title{Convert an entire alignment into a features object}
\usage{
convertAlignmentToFeature(
  aln,
  refseq,
  feature = NULL,
  strand = NULL,
  frame = NULL,
  attribute = NULL
)
}
\arguments{
\item{aln}{msa object representing the alignment}

\item{refseq}{the name of the sequence in the alignment that is used as a frame of reference}

\item{feature}{the feature type name (e.g., "exon", "intron", "CDS", etc. Default NULL)}

\item{strand}{a character string denoting the strand (e.g. "+", "-". Default NULL if strand is not relevant)}

\item{frame}{a 0, 1, or 2 specifying whether the feature is in frame}

\item{attribute}{a character string denoting the label of the feature}
}
\value{
featureout a features object ready to use for phyloConverge
}
\description{
Convert an entire alignment into a features object
}
