% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{convertBedToFeature}
\alias{convertBedToFeature}
\title{Convert a BED matrix/data.frame containing information on genetic elements into a features object}
\usage{
convertBedToFeature(
  bed_file,
  refseq,
  feature = NULL,
  strand = NULL,
  frame = NULL,
  attribute = NULL
)
}
\arguments{
\item{bed_file}{a BED format matrix/data frame containing a list of genetic elements with their chromosome, coordinates, and name (optional)}

\item{refseq}{the name of the sequence in the alignment that is used as a frame of reference}

\item{feature}{the feature type name (e.g., "exon", "intron", "CDS", etc. Default NULL)}

\item{strand}{a character string denoting the strand (e.g. "+", "-". Default NULL if strand is not relevant)}

\item{frame}{a 0, 1, or 2 specifying whether the feature is in frame}

\item{attribute}{a character string denoting the label of the feature}
}
\value{
featureout a features object ready to use for phyloConverge
}
\description{
Convert a BED matrix/data.frame containing information on genetic elements into a features object
}
