% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{getElementsInMaf}
\alias{getElementsInMaf}
\title{Identify which genetic elements in a table exist in a given multiple alignment file}
\usage{
getElementsInMaf(elements_bed, maf)
}
\arguments{
\item{elements_bed}{a BED format matrix/data frame containing a list of genetic elements with their chromosome and coordinates}

\item{maf}{an MSA object containing the sequence alignment}
}
\value{
elements_in_maf a BED format matrix/data frame listing the elements that exist in maf
}
\description{
Identify which genetic elements in a table exist in a given multiple alignment file
}
