% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_tree_processing.R
\name{getForegroundsFromTree}
\alias{getForegroundsFromTree}
\title{Converts a binary phenotype tree to a vector of foreground species names}
\usage{
getForegroundsFromTree(tree)
}
\arguments{
\item{tree}{a phylo object with binary edge lengths (foreground edges = 1, background edges = 0)}
}
\value{
foregrounds a vector of foreground species names
}
\description{
Converts a binary phenotype tree to a vector of foreground species names
}
