% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_tree_processing.R
\name{getInternalBranchName}
\alias{getInternalBranchName}
\title{Produces phyloConverge-ready name of an ancestral branch}
\usage{
getInternalBranchName(node, tree, allEdges, num_tips)
}
\arguments{
\item{node}{internal node ID number, corresponding to the node ID names in the tree object}

\item{tree}{a phylo object with the tree topology of interest}

\item{allEdges}{a matrix containing all edges (e.g., the 'edge' component in the tree object, or a subset of it)}

\item{num_tips}{total number of tip species in the topology}
}
\value{
branch_name the name of the input ancestral node
}
\description{
Produces phyloConverge-ready name of an ancestral branch
}
