% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_tree_processing.R
\name{getTreeFromEvolMod}
\alias{getTreeFromEvolMod}
\title{Returns tree (phylo) object from evolution model}
\usage{
getTreeFromEvolMod(mod)
}
\arguments{
\item{mod}{path to evolution model file}
}
\value{
tree tree object in the evolution model
}
\description{
Returns tree (phylo) object from evolution model
}
