% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_tree_processing.R
\name{getTreeFromForegrounds}
\alias{getTreeFromForegrounds}
\title{Converts foreground species names into a binary phenotype tree with the given topology}
\usage{
getTreeFromForegrounds(foregrounds, topology, plotTree = F)
}
\arguments{
\item{foregrounds}{a character vector of foreground species}

\item{topology}{the intended tree topology represented as a phylo object}

\item{plotTree}{Boolean flag for plotting the resulting binary tree}
}
\value{
treeout binary phenotype tree
}
\description{
Converts foreground species names into a binary phenotype tree with the given topology
}
