% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloConverge.R
\name{phyloConverge}
\alias{phyloConverge}
\title{Computes bias-corrected conservation/acceleration score of a given feature using maximum likelihood estimation of branch scaling and phylogenetic permulations (phylogeny-aware permutation test)}
\usage{
phyloConverge(
  foregrounds,
  permulated_foregrounds,
  neutralMod,
  maf,
  refseq,
  feature = NULL,
  alpha = 0.05,
  min.fg = 2,
  method = "LRT",
  mode = "CONACC",
  adapt = T
)
}
\arguments{
\item{foregrounds}{a character vector containing the names of the foreground species}

\item{permulated_foregrounds}{a list object containing the sets of fake foreground species generated by permulation}

\item{neutralMod}{neutral nucleutide substitution model}

\item{maf}{an MSA object containing the sequence alignment}

\item{refseq}{reference species}

\item{feature}{a feature object containing information on chromosomes, coordinates, and names (optional) of the features to be scored}

\item{alpha}{target significance level to control for}

\item{min.fg}{minimum number of foreground species required}

\item{method}{scoring method for phyloP (default "LRT")}

\item{mode}{scoring mode for phyloP (default "CONACC")}

\item{adapt}{Boolean flag for performing adaptive permulation (adapt = TRUE for adaptive permulation, adapt = FALSE for complete permulations, default TRUE)}
}
\value{
out a data frame containing phyloConverge permulation p-values, corrected score, and uncorrected score (negative score denotes acceleration, positive score denotes deceleration)
}
\description{
Computes bias-corrected conservation/acceleration score of a given feature using maximum likelihood estimation of branch scaling and phylogenetic permulations (phylogeny-aware permutation test)
}
