% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloConverge.R
\name{scanWithPhyloConverge}
\alias{scanWithPhyloConverge}
\title{Scan a given sequence alignment to compute bias-corrected conservation/acceleration scores}
\usage{
scanWithPhyloConverge(
  maf,
  foregrounds,
  permulated_foregrounds,
  neutralMod,
  refseq,
  offset = NULL,
  stride = 1,
  window = 1,
  alpha = 0.05,
  min.fg = 2,
  method = "LRT",
  mode = "CONACC",
  adapt = T
)
}
\arguments{
\item{maf}{an MSA object containing the sequence alignment}

\item{foregrounds}{a character vector containing the names of the foreground species}

\item{permulated_foregrounds}{a list object containing the sets of fake foreground species generated by permulation}

\item{neutralMod}{neutral nucleutide substitution model}

\item{refseq}{reference species}

\item{offset}{offset for the coordinates in the alignment. If NULL, the offset specified in the alignment is used}

\item{stride}{interval (in bp) of positions of sliding windows to be scored (default 1, meaning scoring every nucleotide)}

\item{window}{size (in bp) of the sliding windows to be scored (default 1)}

\item{alpha}{target significance level to control for}

\item{min.fg}{minimum number of foreground species required}

\item{method}{scoring method for phyloP (default "LRT")}

\item{mode}{scoring mode for phyloP (default "CONACC")}

\item{adapt}{Boolean flag for performing adaptive permulation (adapt = TRUE for adaptive permulation, adapt = FALSE for complete permulations, default TRUE)}
}
\value{
out_df a data frame containing the start and end coordinates of the sliding windows and their corresponding phyloConverge scores
}
\description{
Scan a given sequence alignment to compute bias-corrected conservation/acceleration scores
}
