% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permulation_functions.R
\name{getPermulatedPhenotypeTree}
\alias{getPermulatedPhenotypeTree}
\title{Produce a permulated phenotype tree from the observed phenotype and phylogeny}
\usage{
getPermulatedPhenotypeTree(foregrounds, neutraltree, root_species)
}
\arguments{
\item{foregrounds}{a vector of foreground species names}

\item{neutraltree}{a phylo object representing neutral evolution}

\item{root_species}{the species to root the tree on}
}
\value{
permulated_tree a permulated phenotype tree with binary edge lengths (foreground edges = 1, background edges = 0)
}
\description{
Produce a permulated phenotype tree from the observed phenotype and phylogeny
}
